Use dumb argument parser
An expression like `a <|> b :: Parser _` only falls back to `b` if `a`
fails in the "required option missing" sense, and not (for example) in
the "`eitherReader` returned `Left`" sense.

That, as far as I can tell, makes it impossible to do the parser so the
it comes out as `[ LocalPath S3Uri | ... ]` as we'd like, since I can't
make a two-argument local-remote parser "fail" and fall-back into
remote-local, and finally remote-remote, and never attempt to parse
local-local.

Instead, I need to parse `[ LocalPath | S3Uri ]` twice, and even then I
have to put the "or" semantics into the function given to `eitherReader`
to avoid the same issue, which leaves `--help` not-ideal still, but at
least function. And since this style has no way to avoid parsing
local-local as valid, I need to handle that outside, which means `error`
(or not using a `Parser SyncOptions`, which is annoying). Sigh.

It's possible `optparse-applicative` is "wrong" here. In my issue
`envparse`[^1], I'm actually calling how `envparse` behaves a bug citing
`optparse-applicative` as the "right" way. But now I'm having second
thoughts.

[^1]: https://github.com/supki/envparse/issues/23